% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addRawFC}
\alias{addRawFC}
\title{Add more FCs to topTable}
\usage{
addRawFC(tt)
}
\arguments{
\item{tt}{A topTable}
}
\value{
A new topTable with absLogFC and rawFC
}
\description{
Takes a limma topTable as input, and adds the absolute value of the log2 FC, and the raw FC (2^logFC).
}
\examples{
# fit <- lmFit(eset,design)
# fit <- contrasts.fit(fit, contrast.matrix)
# fit <- eBayes(fit)
# tt <- topTable(fit, coef="mycontrast", number=nrow(fit))
# tt <- addRawFC(tt)
}
\keyword{keywords}

