% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.hue.R
\name{adjust.hue}
\alias{adjust.hue}
\title{Adjust colours. degree < 1 means lighter. > 1 means darker}
\usage{
adjust.hue(colours, degree)
}
\description{
Adjust colours. degree < 1 means lighter. > 1 means darker
}
