% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/average_effect.R
\name{adjustinc}
\alias{adjustinc}
\title{adjustinc
this closure gives individual rows of an adjusted grid one at a time, each time it is called.}
\usage{
adjustinc(adjust)
}
\description{
adjustinc
this closure gives individual rows of an adjusted grid one at a time, each time it is called.
}
\examples{
\dontrun{
adjustnext<-adjustinc(expand.grid(b=c(0:5)))
adjustnext()}
}

