% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{aliases}
\alias{aliases}
\alias{brackets}
\alias{dollar}
\alias{is_in}
\alias{and}
\alias{or}
\alias{equals}
\alias{is_gt}
\alias{is_geq}
\alias{is_lt}
\alias{is_leq}
\alias{not}
\alias{set_colnames}
\alias{set_rownames}
\alias{set_names}
\alias{set_class}
\alias{set_attr}
\alias{set_attributes}
\title{Aliases}
\description{
Aliases more pleasing to use in a dplyr pipeline:
\tabular{ll}{
\code{brackets}        \tab \code{`[[`}           \cr
\code{dollar}          \tab \code{`$`}            \cr
\code{is_in}           \tab \code{`\%in\%`}       \cr
\code{and}             \tab \code{`&`}            \cr
\code{or}              \tab \code{`|`}            \cr
\code{equals}          \tab \code{`==`}           \cr
\code{is_gt}           \tab \code{`>`}            \cr
\code{is_geq}          \tab \code{`>=`}           \cr
\code{is_lt}           \tab \code{`<`}            \cr
\code{is_leq}          \tab \code{`<=`}           \cr
\code{not}             \tab \code{`!`}            \cr
\code{set_colnames}    \tab \code{`colnames<-`}   \cr
\code{set_rownames}    \tab \code{`rownames<-`}   \cr
\code{set_names}       \tab \code{`names<-`}      \cr
\code{set_attributes}  \tab \code{`attributes<-`} \cr
\code{set_attr}        \tab \code{`attr<-`}       \cr
}
}
