% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quartet.R
\docType{data}
\name{anscombe}
\alias{anscombe}
\title{Anscombe's Quartet data}
\format{Data frame with columns}
\source{
Anscombe's Quartet,
  \url{http://en.wikipedia.org/wiki/Anscombe\%27s_quartet}
}
\usage{
anscombe
}
\description{
Tidy version of built-in Anscombe's Quartet data. Four datasets that have
nearly identical linear regression properties, yet appear very different when
graphed.
}
\keyword{datasets}

