% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.results.R
\name{append.results}
\alias{append.results}
\title{Function that appends to lists/data.frame instead of overwriting them
(in contrast to load())}
\usage{
append.results(file, add.filename = FALSE, ...)
}
\description{
Function that appends to lists/data.frame instead of overwriting them
(in contrast to load())
}
