% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_all_equal}
\alias{are_all_equal}
\title{Are all equal?}
\usage{
are_all_equal(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Remove missing values (FALSE by default; NAs in x will return NA).}
}
\value{
Logical, whether all elements of a numeric vector are equal.
}
\description{
Are all the elements of a numeric vector (approximately) equal?
}
\examples{
are_all_equal(c(5,5,5))
are_all_equal(c(5,5,5,6))
are_all_equal(c(5,5,5,NA,6))
are_all_equal(c(5,5,5,NA,6), na.rm=TRUE)
5==5.000000001
identical(5, 5.000000001)
are_all_equal(c(5L, 5, 5.000000001))

}
