% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beep.R
\name{beep}
\alias{beep}
\title{Play a short sound}
\usage{
beep(expr = NULL)
}
\arguments{
\item{expr}{An optional expression to be excecuted before the sound.}
}
\description{
\code{beep} plays a short sound which is useful if you want to get notified, 
for example, when a script has finished. Taken from the \code{beepr} package.
}
\details{
If \code{beep} is not able to play the sound a warning is issued rather than 
an error. This is in order to not risk aborting or stopping the process that
you wanted to get notified about.
}
\examples{
\dontrun{
# Play a beep sound
beep()


# Update all packages and "ping" when it's ready
update.packages(ask=FALSE); beep()
}
}

