% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxcartry.R
\name{boxcartry}
\alias{boxcartry}
\title{boxcartry}
\usage{
boxcartry(token = .boxcar_token, taskname = "task", ...)
}
\arguments{
\item{token}{the boxcar token}

\item{taskname}{the name of the task to be run.}
}
\description{
Execute an expression, sending a boxcar notification when done or on error.
}

