% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_rda.R
\name{convertcsv}
\alias{convertcsv}
\title{convertcsv}
\usage{
convertcsv(dir = "./", save_format = "rda")
}
\description{
converts files in directory from dta to rda and saves copies.
Among other things, this is a useful function to use if you need #' or want to save some harddrive space.
}

