% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_rda.R
\name{convertxls}
\alias{convertxls}
\title{convertxls}
\usage{
convertxls(mydir = "./")
}
\description{
converts files in directory from xls to rda and saves copies.
}
\note{
You should/must put a trailing slash on the end of the path you are searching.
}

