% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cs.R
\name{Cs}
\alias{Cs}
\title{Character strings from unquoted names}
\usage{
Cs(...)
}
\arguments{
\item{...}{any number of names separated by commas}
}
\value{
character string vector
}
\description{
Makes a vector of character strings from a list of valid names. Taken from Hmisc.
}
\examples{
Cs(a, cat, dog)
Cs(age, sex, race, bloodpressure, height)
Cs(you, must, quote, "things with spaces", 'or invalid ! @#$ characters')

}
\author{
Frank Harrell
}
