\name{datename}
\alias{datename}
\title{Filename with date}
\usage{
datename(filename = "filename")
}
\arguments{
  \item{filename}{A filename string}
}
\value{
String with today's date in YYYY-MM-DD- format concatenated
to filename.
}
\description{
Returns string with today's date in YYYY-MM-DD- format
concatenated to filename.
}
\examples{
datename("results.csv")
}
\keyword{keywords}

