% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.copy2PDF.R
\name{dev.copy2PDF}
\alias{dev.copy2PDF}
\title{overrides default to enable similar behaviour from interactive and
non-interactive use
used in conjunction with PDF()}
\usage{
dev.copy2PDF(file, ..., width = NULL, height = NULL)
}
\description{
overrides default to enable similar behaviour from interactive and
non-interactive use
used in conjunction with PDF()
}
