% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dfclass.R
\name{dfclass}
\alias{dfclass}
\title{Get class of all data frame columns}
\usage{
dfclass(df)
}
\arguments{
\item{df}{A data.frame object}
}
\value{
A data.frame with index and class
}
\description{
Get class of all data frame columns in a friendly format
}
\examples{
dfclass(iris)
}
\author{
Stephen Turner
}
\keyword{keywords}

