% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfsigfig.R
\name{dfsigfig}
\alias{dfsigfig}
\title{Round numeric columns of a data frame}
\usage{
dfsigfig(df, n = 3)
}
\arguments{
\item{df}{A data.frame.}

\item{n}{The number of significant digits to round off to.}
}
\value{
A data.frame rounded to n significant digits.
}
\description{
Round the numeric columns of a data frame to a specified number of significant digits.
}
\examples{
\dontrun{
dfsigfig(mtcars,1)
}

}
\author{
Stephen Turner
}
