% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.from.RData.R
\name{extract.from.RData}
\alias{extract.from.RData}
\title{Function for extracting an object from a .RData file created by R's save() command}
\usage{
extract.from.RData(file, name)
}
\description{
Function for extracting an object from a .RData file created by R's save() command
}
