% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2char.R
\name{factor2char}
\alias{factor2char}
\title{Function to turn factor variables in df (data.frame) to characters}
\usage{
factor2char(df)
}
\description{
Function to turn factor variables in df (data.frame) to characters
}

