% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpos.equal.R
\name{findpos.equal}
\alias{findpos.equal}
\title{Basically findpos(input == within) but maintain the order of input}
\usage{
findpos.equal(input, within, firstonly = FALSE, lastonly = FALSE)
}
\description{
Basically findpos(input == within) but maintain the order of input
}
