% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpos.R
\name{findpos}
\alias{findpos}
\title{Function to find position of TRUEs in a logical vector. 
You can optionally put in a "within" vector to find the positions
within the "within" vector.}
\usage{
findpos(vec, within = NULL, firstonly = FALSE, lastonly = FALSE)
}
\arguments{
\item{vec}{A logical vector or a findpos object}

\item{within}{A logical vector or a findpos object}
}
\description{
Function to find position of TRUEs in a logical vector. 
You can optionally put in a "within" vector to find the positions
within the "within" vector.
}
