% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fisherp}
\alias{fisherp}
\title{Fisher's method to combine p-values.}
\usage{
fisherp(x)
}
\arguments{
\item{x}{A vector of p-values between 0 and 1.}
}
\value{
A combined p-value.
}
\description{
Uses Fisher's method to combine p-values from different tests.
}
\examples{
fisherp(c(.042, .02, .001, 0.01, .89))

}
