% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get1000GpseudoID.R
\name{get1000GpseudoID}
\alias{get1000GpseudoID}
\title{Find the pseudoID used in the 1000G dataset on our server 
given rsID or chromosome and position and alleles}
\usage{
get1000GpseudoID(df, rm.duplicates = T, chromosomes = c(1:22, "X", "Y",
  "M"), cluster = NULL, ...)
}
\description{
Find the pseudoID used in the 1000G dataset on our server 
given rsID or chromosome and position and alleles
}
