% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_na.R
\name{gg_na}
\alias{gg_na}
\title{Plot missing data}
\usage{
gg_na(df)
}
\arguments{
\item{df}{A data.frame.}
}
\description{
Plots missing data as holes on a black canvas.
}
\examples{
set.seed(2016-07-12)
x <- matrix(1, nrow=50, ncol=50)
x[sample(prod(dim(x)), 100)] <- NA
x <- data.frame(x)
gg_na(x)

}
\author{
Stephen Turner
}
\keyword{NA}

