% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{gghues}
\alias{gghues}
\title{Emulate ggplot2 default hues}
\usage{
gghues(n, start = 15)
}
\arguments{
\item{n}{The Numeric; number of hues to generate.}

\item{start}{Numeric; the place on the color wheel to start. ggplot2 default is 15.}
}
\value{
A vector of hues
}
\description{
This will emulate ggplot2's hues, which are equally spaced hues around the color wheel, starting from 15.
}
\examples{
n <- 10
gghues(3)
barplot(rep(1,n), col=gghues(n), names=gghues(n))
barplot(rep(1,n), col=gghues(n, start=15+180), names=gghues(n, start=15+180))

}
