% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grapes.R
\name{\%like\%}
\alias{\%like\%}
\title{x like y}
\usage{
x \%like\% pattern
}
\arguments{
\item{x}{a vector (numeric, character, factor)}

\item{pattern}{a vector (numeric, character, factor), matching the mode of x}
}
\value{
A logical vector with length equal to \code{x} of things in \code{x} that are like \code{y}.
}
\description{
Returns a logical vector of elements of x matching the regex y.
}
\examples{
(Name <- c("Mary","George","Martha"))
Name \%in\% c("Mary")
Name \%like\% "^Mar"
Name \%nin\% c("George")
Name \%nlike\% "^Mar"
}
\seealso{
\code{\link{\%like\%}}, \code{\link{\%nlike\%}}, \code{\link{\%nin\%}},
}
