% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grapes.R
\name{\%+\%}
\alias{\%+\%}
\title{String concatenation}
\usage{
x \%+\% y
}
\arguments{
\item{x}{a string}

\item{y}{a string}
}
\value{
A concatenated string
}
\description{
Infix operator for [paste0].
}
\examples{
paste0("Hello, ", "this package name is ", "Tmisc")
"Hello, " \%+\% "this package name is " \%+\% "Tmisc!"
cat("This is the number " \%+\% seq(1,10) \%+\% "\n")
}
