% Generated by roxygen2 (4.0.1): do not edit by hand
\name{h}
\alias{h}
\title{Return the First or Last Part of an Object}
\usage{
h(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{other arguments to be passed to \code{head}}
}
\value{
An object like x but smaller
}
\description{
Shortcut to \code{\link{head}}.
}
\examples{
h(iris)
}
\author{
Stephen Turner
}
\keyword{keywords}

