% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertInAddin.R
\name{insertInAddin}
\alias{insertInAddin}
\title{Insert \%in\%.}
\usage{
insertInAddin()
}
\description{
Call this function as an addin to insert \code{ \%in\% } at the cursor
position. After installing Tmisc, hit the Addins menu, and optionally add a
keyboard shortcut, e.g., Command-Shift-I.
}
