% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.packages.R
\name{install.packages}
\alias{install.packages}
\title{Function to install to local MRAN library if from MRAN. 
Otherwise install to default (should be attachroot("/WORK/Rpackages/nonMRAN"))}
\usage{
install.packages(..., repos = getOption("repos"))
}
\description{
Function to install to local MRAN library if from MRAN. 
Otherwise install to default (should be attachroot("/WORK/Rpackages/nonMRAN"))
}
