% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ipak}
\alias{ipak}
\title{Install packages}
\usage{
ipak(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{install.packages}}
}
\description{
Shortcut to \code{\link{install.packages}}
}
\examples{
\dontrun{
lsp(Tmisc, pattern="un")
}
}
\author{
Stephen Turner
}
\keyword{keywords}

