% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.results.R
\name{load.results}
\alias{load.results}
\title{Load a bunch of .RData files and have them append.results() together}
\usage{
load.results(pattern, path = ".", ext = ".RData", prompt = T,
  append = F)
}
\description{
Load a bunch of .RData files and have them append.results() together
}

