% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuszoom.R
\name{locuszoom}
\alias{locuszoom}
\title{Wrapper around locuszoom (1.4)}
\usage{
locuszoom(data = NULL, ..., type = c("auto", "metal", "epacts"),
  chr.col = NULL, pos.col = NULL, marker.col = NULL, pval.col = NULL,
  annot.col = NULL, ld = NULL,
  locuszoom.dir = "~/software/locuszoom_tim/", ld.cache = tempfile(),
  Rplink.opts = list(), plot = TRUE, verbose = debug, debug = FALSE,
  help = FALSE, keepfiles = FALSE)
}
\details{
Note that the conf/m2zfast.conf file in locuszoom must be set up to point to the various
dependent programs first. The github version of locuszoom (statgen/locuszoom-standalone) appears to be missing
the conf/m2zfast.conf file as of 19/4/2018. If that's the case, simply clone a copy of
conf/dist.conf as conf/m2zfast.conf, and modify the various variables.
}
