% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lsa}
\alias{lsa}
\title{List objects and classes}
\usage{
lsa()
}
\value{
A data frame with all objects and their class.
}
\description{
List all the objects and their classes in the global environment.
}
\examples{
# lsa()
}
\keyword{keywords}

