% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergett.R
\name{mergett}
\alias{mergett}
\title{Merge topTable with expression values}
\usage{
mergett(tt, eset)
}
\arguments{
\item{tt}{A topTable}

\item{eset}{An ExpressionSet object from which tt derives}
}
\value{
An expanded topTable with gene expression values merged in, sorted by
  p-value.
}
\description{
Merges topTable output with normalized expression values from the
ExpressionSet. Note: run \code{addRawFC(tt)} before merging with
ExpressionSet values.
}
\examples{
\dontrun{
fit <- lmFit(eset,design)
fit <- contrasts.fit(fit, contrast.matrix)
fit <- eBayes(fit)
tt <- topTable(fit, coef="mycontrast", number=nrow(fit))
tt <- addRawFC(tt)
tt <- mergett(tt, eset)
}
}
\author{
Stephen Turner
}
\keyword{keywords}

