% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Function to get min and max for a bunch of objects (all coerced to vectors)
to set xlim and ylim in plot}
\usage{
minmax(..., na.rm = TRUE)
}
\description{
Function to get min and max for a bunch of objects (all coerced to vectors)
to set xlim and ylim in plot
}
