% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode.}
\usage{
Mode(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Remove missing values before calculating the mode (TRUE by default).}
}
\value{
A combined p-value.
}
\description{
Returns the mode of a vector. First in a tie wins (see examples).
}
\examples{
Mode(c(1,2,2,3,3,3,10))
Mode(c("A", "C", "C", "B", "B"))
Mode(c(10,20,20,30,30,30,NA, NA, NA, NA, NA, 100))
Mode(c(1,2,2,3,3,3,NA, NA, NA, NA, NA, 10), na.rm=FALSE)

}
\author{
Stephen Turner
}
\keyword{keywords}

