% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.last.cols.R
\name{move.last.cols}
\alias{move.last.cols}
\title{Move the last n columns to the middle of the data.frame/matrix}
\usage{
move.last.cols(df, n, to)
}
\description{
Move the last n columns to the middle of the data.frame/matrix
}
