% Generated by roxygen2 (4.0.1): do not edit by hand
\name{n}
\alias{n}
\title{Numbered list of column names}
\usage{
n(df)
}
\arguments{
\item{df}{A data.frame object}
}
\value{
A single column numbered list of data.frame names
}
\description{
Given data.frame df, return the `names` of that data.frame
in a "long" one-column numbered format.
}
\examples{
n(iris)
}
\author{
Stephen Turner
}
\keyword{keywords}

