% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameoption.R
\name{nameoption}
\alias{nameoption}
\title{Function to extract option from a filename
We assume the filename will contain information about
options used in a simiulation, e.g. sim.10.DOM.R
nameoption("sim.10.DOM.R", 2) --> 10
nameoption("sim.10.DOM.R", 3) --> "DOM"}
\usage{
nameoption(filename, pos, split = ".", ...)
}
\description{
Function to extract option from a filename
We assume the filename will contain information about
options used in a simiulation, e.g. sim.10.DOM.R
nameoption("sim.10.DOM.R", 2) --> 10
nameoption("sim.10.DOM.R", 3) --> "DOM"
}

