% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\name{neighbours}
\alias{neighbours}
\title{Returns a sequence of number centred at a number i}
\usage{
neighbours(i, plusminus = 3, min = 1, max = Inf)
}
\description{
Returns a sequence of number centred at a number i
}
