% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nn.R
\name{nn}
\alias{nn}
\title{Numbered list of column names}
\usage{
nn(df)
}
\arguments{
\item{df}{A data.frame object}
}
\value{
A single column numbered list of data.frame names
}
\description{
Given data.frame df, return the `names` of that data.frame
in a "long" one-column numbered format.
}
\examples{
nn(iris)
}
\author{
Stephen Turner
}
\keyword{keywords}

