% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDF.R
\name{PDF}
\alias{PDF}
\title{start pdf if non-interactive
saves some global variable to be used by dev.copy2pdf()}
\usage{
PDF(file, ..., width = 7, height = 7)
}
\description{
start pdf if non-interactive
saves some global variable to be used by dev.copy2pdf()
}
