% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{peek}
\alias{peek}
\title{Peek at the top of a text file}
\usage{
peek(x, n = 5)
}
\arguments{
\item{x}{a filename}

\item{n}{the number of lines to return}
}
\description{
This returns a character vector which shows the top n lines of a file.
}
\examples{
\dontrun{
filename <- tempfile()
x<-matrix(round(rnorm(10^4),2),1000,10)
colnames(x)=letters[1:10]
write.csv(x,file=filename,row.names=FALSE)
peek(filename)
}
}
