% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.version.R
\name{pkg.version}
\alias{pkg.version}
\title{Find out the package version of specified packages
Not to be confused with base::package_version()
but similar to utils::packageVersion()}
\usage{
pkg.version(pkg.vec)
}
\description{
Find out the package version of specified packages
Not to be confused with base::package_version()
but similar to utils::packageVersion()
}
