% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.results.table.R
\name{plot.results.table}
\alias{plot.results.table}
\title{Automatic plotting of results.table...}
\usage{
\method{plot}{results.table}(data, x = NULL, y = NULL, fill = NULL,
  X = NULL, Y = NULL, wrap = NULL, scales = NULL)
}
\description{
Automatic plotting of results.table...
}
