\name{propmiss}
\alias{propmiss}
\title{Missing stats}
\usage{
propmiss(df)
}
\arguments{
  \item{df}{A data.frame object}
}
\value{
A data.frame with missingness stats
}
\description{
Returns the number of missing values, total length, and
proportion missing values for each variable in a data.frame
}
\examples{
propmiss(iris)
}
\keyword{keywords}

