% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quartet}
\alias{quartet}
\title{Anscombe's Quartet data (tidy)}
\format{
Data frame with three columns, set, x, y.
}
\usage{
quartet
}
\description{
Tidy version of built-in Anscombe's Quartet data. Four datasets that have
nearly identical linear regression properties, yet appear very different when
graphed.
}
\keyword{datasets}
