% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.gist}
\alias{read.gist}
\title{Read data from a github gist}
\usage{
read.gist(gistid, ...)
}
\arguments{
\item{gistid}{A GitHub gist ID.}

\item{...}{Other arguments passed to \code{read.table()}}
}
\value{
A data frame with the data you read in from the github gist.
}
\description{
Read data from a github gist. Modified from Hadley Wickham's
\code{source_gist()} function in devtools. Call the function with the gist
ID. E.g., if url is \url{https://gist.github.com/user/abc123}, the ID is
abc123.
}
\examples{
# read.gist("6fa823bb2de7e541ad54", header=TRUE)
}
\keyword{keywords}

