% Generated by roxygen2 (4.0.1): do not edit by hand
\name{registration}
\alias{registration}
\title{Get registration link}
\usage{
registration(x = NULL)
}
\arguments{
\item{x}{Keep \code{NULL} for now.}
}
\value{
A registration link
}
\description{
Returns a link to register for a course. This is used to ensure that
participants were able to successfully download and install R and Rstudio
before coming to the course.
}
\examples{
registration()
}
\author{
Stephen Turner
}
\keyword{keywords}

