% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.fread.R
\name{remote.fread}
\alias{remote.fread}
\title{Function to fread directly from server}
\usage{
remote.fread(file, ssh.server = "hpcf2.cgs.hku.hk", cat = TRUE, ...)
}
\description{
Function to fread directly from server
}
