% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repglm.lrt.R
\name{repglm.lrt}
\alias{repglm.lrt}
\title{Function to perform Likelihood Ratio test, given two lists of deviances and their
df as output from repglm.wrapper(..., type="deviance")}
\usage{
repglm.lrt(dev1, dev0)
}
\description{
Function to perform Likelihood Ratio test, given two lists of deviances and their
df as output from repglm.wrapper(..., type="deviance")
}
