% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale a vector}
\usage{
rescale(x, to = c(0, 1))
}
\arguments{
\item{x}{A numeric vector to rescale.}

\item{to}{output range (numeric vector of length two)}
}
\value{
The numeric vector scaled from zero to one.
}
\description{
Rescale a vector to have specified min and max. Borrowed from the scales package.
}
\examples{
rescale(runif(10))
rescale(1:11)
rescale(c(1,1))
rescale(1)
rescale(c(1, NA, 2))

}
\author{
Stephen Turner
}
\keyword{NA}

