% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale01.R
\name{rescale01}
\alias{rescale01}
\title{Rescale 0 to 1}
\usage{
rescale01(x)
}
\arguments{
\item{x}{A numeric vector to rescale.}
}
\value{
The numeric vector scaled from zero to one.
}
\description{
Rescales a vector from 0 to 1. Borrowed from the scales package.
}
\examples{
rescale01(runif(10))
rescale01(1:11)
rescale01(c(1,1))
rescale01(1)
rescale01(c(1, NA, 2))

}
\author{
Stephen Turner
}
\keyword{NA}

