% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio.R
\name{revcomp}
\alias{revcomp}
\title{Reverse complement}
\usage{
revcomp(x)
}
\arguments{
\item{x}{A sequence to reverse complement}
}
\value{
The sequence, reverse complemented
}
\description{
Reverse complements a sequence.
}
\examples{
revcomp("GATTACA")
sapply(c("GATTACA", "CATATTAC"), revcomp)

}
