% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmixture.normal.R
\name{rmixture.normal}
\alias{rmixture.normal}
\title{Generate random mixure Normal variables
Mixture of 0 and rnorm(n.causal, ...)}
\usage{
rmixture.normal(n, p.causal, ...)
}
\description{
Generate random mixure Normal variables
Mixture of 0 and rnorm(n.causal, ...)
}
