% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/safesave.R
\name{safesave}
\alias{safesave}
\title{safesave}
\usage{
safesave(...)
}
\arguments{
\item{...}{Arguments passed to `save()`}
}
\value{
prints message indicating whether save occured or how to set options to enable save
}
\description{
A function that only saves when options('saveopts') is set to TRUE
}
\details{

}
\examples{
\dontrun{
x<-rnorm(100)
options('saveopts'=F)
safesave(x,file='./tmp.Rda')
# will not save file
options('saveopts'=T)
safesave(x,file='./tmp.Rda')
# will save

# use case: setup a script and source, when you want to save it, set options before sourcing:
# script.R
x<-rnorm(100)
y<-runif(199)
z<-rpois(100)
safesave(x,file='./tmp.Rda')
safesave(y,file='./tmp1.Rda')
safesave(z,file='./tmp2.Rda')

#---
options('saveopts'=F) # default,
source('./script.R') # will run but not save

# Script can save up to 3 files, but we don't always want it to do that. By default no files will save. to save (ex. from bash):
options('saveopts'=T)
source('./script.R') # will save all three files.

# alternative in bash
# $ Rscript -e 'options("saveopts"=T);source("file.R")'
}
}

