% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sicb}
\alias{sicb}
\title{Write sessionInfo to the clipboard}
\usage{
sicb()
}
\description{
Writes output of \code{sessionInfo()} to the clipboard. Only works on Mac.
}
\examples{
\dontrun{
# Write sessionInfo() to the clipboard on mac.
sicb()
}

}
