% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slack_notify.R
\name{slack_notify}
\alias{slack_notify}
\title{slack_notify}
\usage{
boxcar_notify(token,body,title)
}
\arguments{
\item{webhookuri}{this is the uri that the webhook integration of slack will give you for posting into a specified channel. You need to configure this through slack.}

\item{title}{the title of the message (this appears first and foremost in the message.)}

\item{body}{the text of the body of the warning. A character string.}
}
\value{
Sends a push notification to the specified team chat  via slack.
}
\description{
A simple function that sends a push notification via slack
}
\details{

}
\examples{
\dontrun{
#.slack_webhook_uri<-"https://hooks.slack.com/services/xxxxx/xxxxx/xxxxx"
slack_notify(webhookuri=.slack_webhook_uri,title="this is an example","example")
}
}

