% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.some.R
\name{source.some}
\alias{source.some}
\title{Function to source part of an R script. 
Helps avoid duplication of codes!}
\usage{
source.some(file, start = "", end = "", run = FALSE, ...)
}
\description{
Function to source part of an R script. 
Helps avoid duplication of codes!
}
