% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr.R
\name{strr}
\alias{strr}
\title{Function to examine the object like str
but will try to avoid giving displays that's too long}
\usage{
strr(object, nmax = 10, tail = F)
}
\description{
Function to examine the object like str
but will try to avoid giving displays that's too long
}
