% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strSort.R
\name{strSort}
\alias{strSort}
\title{Sort characters in a string}
\usage{
strSort(x)
}
\arguments{
\item{x}{A string to sort.}
}
\value{
A sorted string.
}
\description{
Alphabetically sorts characters in a string. Vectorized over x.
}
\examples{
strSort("cba")
strSort("zyxcCbB105.a")
strSort(c("cba", "zyx"))
strSort(c("cba", NA))

}
\author{
Stephen Turner
}
\keyword{keywords}
