% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_and_load.R
\name{test_and_load}
\alias{test_and_load}
\title{test_and_load}
\usage{
test_and_load(file_path, object = "poast_dyadic")
}
\arguments{
\item{file_path}{String or file.path() for the path to the file to be loaded if the file is not present.}

\item{object}{The quoted name of the object to be checked.}
}
\description{
A really simple function that checks to see if a file is loaded in the global namespace and then loads from a file if needed
This is useful for anticipating and preventing errors in scripts.
}

