% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_build.r
\name{testObject}
\alias{testObject}
\title{testObject function}
\usage{
testObject(object)
}
\arguments{
\item{object}{This is an object in the current environment (typically the global environment)}
}
\value{
boolean. True if exists, false otherwise
}
\description{
testObject function
}

