% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimStartup.R
\name{TimStartup}
\alias{TimStartup}
\title{Copy this to the start of a script to enable other CGS users to run my scripts}
\usage{
TimStartup()
}
\description{
Copy this to the start of a script to enable other CGS users to run my scripts
}
