% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tmisc-package.R
\docType{package}
\name{Tmisc-package}
\alias{Tmisc}
\alias{Tmisc-package}
\title{Tmisc: Turner Miscellaneous}
\description{
Miscellaneous utility functions for data manipulation, data tidying, and working with gene expression data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenturner/Tmisc}
  \item \url{https://stephenturner.github.io/Tmisc/}
}

}
\author{
\strong{Maintainer}: Stephen Turner \email{vustephen@gmail.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

}
\keyword{internal}
