% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tpairs.R
\name{Tpairs}
\alias{Tpairs}
\title{Better scatterplot matrices.}
\usage{
Tpairs(x, histogram = TRUE, gap = 0, ...)
}
\arguments{
\item{x}{A numeric matrix or data.frame.}

\item{histogram}{Overlay a histogram on the diagonals?}

\item{gap}{distance between subplots, in margin lines.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
A matrix of scatter plots with rugged histograms, correlations, and significance stars. Much of the functionality borrowed from \code{PerformanceAnalytics::chart.Correlation()}.
}
\examples{
Tpairs(iris[-5])
Tpairs(iris[-5], pch=21, bg=Tcols[factor(iris$Species)])
Tpairs(iris[-5], pch=21, bg=gghues(3)[factor(iris$Species)], gap=1)

}
\author{
Stephen Turner
}
\keyword{NA}

