% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.trailing.bases.R
\name{trim.trailing.bases}
\alias{trim.trailing.bases}
\title{If A1 == "GA", and A2 == "GAA", then this is equivalent to 
A1 == "G", A2 == "GA".}
\usage{
trim.trailing.bases(A1, A2, keep.identical = TRUE)
}
\description{
If A1 == "GA", and A2 == "GAA", then this is equivalent to 
A1 == "G", A2 == "GA".
}
