% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unfactor}
\alias{unfactor}
\title{Change factors to characters}
\usage{
unfactor(df)
}
\arguments{
\item{df}{The dataframe you wish to change the factors into characters}
}
\value{
A data.frame with factors converted to characters
}
\description{
Change factors to characters in a dataframe
}
\examples{
df <- data.frame(a = letters[1:5], x = 1:5, y = LETTERS[1:5], stringsAsFactors = TRUE)
str(df)
df <- unfactor(df)
str(df)
}
\author{
\url{https://github.com/Dasonk}
}
\keyword{keywords}

