% Generated by roxygen2 (4.0.0): do not edit by hand
\name{unrowname}
\alias{unrowname}
\title{Un-rowname}
\usage{
unrowname(x)
}
\arguments{
\item{x}{data frame}
}
\value{
Character vector of the gene symbol with the probe ID
}
\description{
Strip rownames from an object (stolen from plyr).
}
\examples{
unrowname(mtcars)
}
\keyword{keywords}

