% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within.range.R
\name{within.range}
\alias{within.range}
\title{Function to determine if x falls within a range}
\usage{
\method{within}{range}(x, x2 = NULL, lower, upper, inclusive = TRUE,
  OR = TRUE, reference = FALSE, no.warning = FALSE)
}
\description{
Function to determine if x falls within a range
}
