% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within.window.R
\name{within.window}
\alias{within.window}
\title{Returns a logical vector indicating if x falls within a 
window of center.}
\usage{
\method{within}{window}(x, center, window, inclusive = TRUE)
}
\description{
Returns a logical vector indicating if x falls within a 
window of center.
}
