% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/average_effect.R
\name{avg_pred_qi}
\alias{avg_pred_qi}
\title{average predicted qi simulation}
\usage{
To use, input a model, the number of simulations you'd like to run, and the features (i.e. variables ) you'd like to hold constant.
}
\description{
This generates the average predicted effect for quantities of interest using simulations. This is useful for examining the effect of a quantity of interest (iv) in an interaction while still remaining sensitive to the dependence of the interaction effect \eqn{\partial^2 E[y|XB]/\partial x_1 \partial x_2}.
}
\examples{
\dontrun{
x<-runif(500,0,1)
z<-runif(500,0,1)
y<-rbinom(500,1,.1*x+.1*z+.5*x*z)
mod1<-glm(y~x*z)
qi<-avg_pred_qi(mod1,500,x=c(0,1))
qi<-avg_pred_qi(mod1,500,x=seq(0,1,.2),z=1)
}
NULL
}

