% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_figure.R
\name{bootstrap_figure}
\alias{bootstrap_figure}
\title{bootstrap_figure}
\usage{
bootstrap_figure(fig_title="",fig_number=NULL,fig_prefix="Figure",content=...)
}
\arguments{
\item{fig_title}{The name of the figure}

\item{fig_number}{the number of the figure. Use a counter(see example) here if you want to update the figure numbers throughout the document}

\item{fig_prefix}{The word/s before the number. Default is "Figure", but you might change it to "number" or whatever}

\item{content}{This is the part that is in the body of the figure. Ideally, it'd be something like an image.}
}
\description{
A block that puts an image into a panel in bootstrap.
}
\note{
The command will throw in NULL for content if left blank

ggplot2 will work with this function, but your expression MUST be enclosed in a print command. See the examples.
}
\examples{
# Setup a counter to use for this
counter<-function(){i<-0;function(){i<<-i+1;return(i)}}
fig_num<-counter()
# run bootstrap_figure on a picture of a historgram
bootstrap_figure(fig_title="title",fig_number="a",content=hist(runif(10,0,1)))
# To run with ggplot, be sure to enclose the expression in a print() command:
require(ggplot2)
bootstrap_figure(fig_title="title",fig_number="a",content=print(qplot(rnorm(1000,0,1),binwidth=.2)))
}

