% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datename.R
\name{datename}
\alias{datename}
\title{Filename with date}
\usage{
datename(filename = "filename")
}
\arguments{
\item{filename}{A filename string.}
}
\value{
String with today's date in YYYY-MM-DD- format concatenated to filename.
}
\description{
Returns string with today's date in YYYY-MM-DD- format concatenated to filename.
}
\examples{
datename("myfile.png")

}
\author{
Stephen Turner
}
\keyword{keywords}
