% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipses.R
\name{ellipses}
\alias{ellipses}
\title{Truncate a data frame with ellipses.}
\usage{
ellipses(df, n = 5L)
}
\arguments{
\item{df}{A data.frame.}

\item{n}{The number of rows to show before an ellipses row.}
}
\value{
A data frame truncated by a row of ellipses.
}
\description{
Prints the specified number of rows of a data frame, followed by a row of ellipses. Useful for piping to \code{knitr::kable()} for printing a truncated table in a markdown document.
}
\examples{
ellipses(mtcars, 5)

}
\author{
Stephen Turner
}

