% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ht.R
\name{ht}
\alias{ht}
\title{Head-tail}
\usage{
ht(d, ...)
}
\arguments{
\item{d}{The data frame or matrix you want to head-tail}

\item{...}{additional arguments passed to head and tail}
}
\value{
The first and last few lines of a data frame or matrix.
}
\description{
Prints the \code{head} and \code{tail} of a data.frame or matrix.
}
\examples{
\dontrun{
ht(iris)
}
}
\author{
Stephen Turner
}
\keyword{keywords}

