\name{ht}
\alias{ht}
\title{Head-tail}
\usage{
ht(d, ...)
}
\arguments{
  \item{d}{The data frame or matrix you want to head-tail}

  \item{...}{additional arguments passed to head and tail}
}
\value{
The first and last few lines of a data frame or matrix.
}
\description{
Prints the \code{head} and \code{tail} of a data.frame or
matrix.
}
\examples{
ht(iris)
}
\keyword{keywords}

