% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsp.R
\name{lsp}
\alias{lsp}
\title{List objects in package}
\usage{
lsp(package, ...)
}
\arguments{
\item{package}{The name of the package you're examining.}

\item{...}{further arguments to be passed to \code{lsf.str}.}
}
\value{
A list of functions and how to call them for any package.
}
\description{
Lists functions and how to call them for any package.
}
\examples{
\dontrun{
lsp(Tmisc, pattern="un")
}
}
\author{
Stephen Turner
}
\keyword{keywords}

