% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode.}
\usage{
Mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Remove missing values before calculating the mode (FALSE by
default). NAs are counted just like any other element. That is, an NA in
the vector won't necessarily result in a return NA. See the first example.}
}
\value{
A combined p-value.
}
\description{
Returns the mode of a vector. First in a tie wins (see examples).
}
\examples{
Mode(c(1,2,2,3,3,3, NA))
Mode(c(1,2,2,3,3,3, NA), na.rm=TRUE)
Mode(c(1,2,2,3,3,3, NA, NA, NA, NA))
Mode(c(1,2,2,3,3,3, NA, NA, NA, NA), na.rm=TRUE)
Mode(c("A", "Z", "Z", "B", "B"))

}
\author{
Stephen Turner
}
\keyword{keywords}
