\name{n}
\alias{n}
\title{Returns names(df) in single column, numbered matrix format.}
\usage{
n(df)
}
\arguments{
  \item{df}{A data.frame object \code{df}}
}
\value{
A single column numbered list of data.frame names
}
\description{
Given data.frame df, return the `names` of that data.frame
in a "long" one-column numbered format.
}
\examples{
n(iris)
}
\keyword{keywords}

