% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{nn}
\alias{nn}
\title{Get names and class of all columns in a data frame}
\usage{
nn(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A data.frame with index and class.
}
\description{
Get names and class of all columns in a data frame in a friendly format.
}
\examples{
nn(iris)

}
\author{
Stephen Turner
}
\keyword{NA}

