\name{ppc}
\alias{ppc}
\title{ppc}
\usage{
bootstrap_figure(fig_title="",fig_number=NULL,fig_prefix="Figure",content=...)
}
\arguments{
  \item{model}{The name of the figure}
}
\description{
Gives ppc and pre for model
}
\note{
Model must be of type 'glm' (or inherit it) for this to
work correclty
}

