% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propmiss.R
\name{propmiss}
\alias{propmiss}
\title{Missing stats}
\usage{
propmiss(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A data.frame with missingness stats.
}
\description{
Returns the number of missing values, total length, and proportion
missing values for each variable in a data.frame
}
\examples{
propmiss(data.frame(a=1:5, b=c(6,NA,NA,9,10)))

}
\author{
Stephen Turner
}
\keyword{NA}
