% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.cb.R
\name{read.cb}
\alias{read.cb}
\title{Read from the clipboard}
\usage{
read.cb(header = TRUE, ...)
}
\arguments{
\item{header}{A logical value indicating whether the file contains the names of the variables as its first line. Overrides the default \code{header=FALSE} option in \code{read.table()}.}

\item{...}{Further arguments to be passed to \code{read.table}}
}
\value{
A data.frame
}
\description{
Read tabular data from the clipboard.
}
\examples{
\dontrun{
# To read CSV data with a header from the clipboard:
read.cb(header=TRUE, sep=',')
}

}
\author{
Stephen Turner
}
\keyword{keywords}
