% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{registration}
\alias{registration}
\title{Get registration link}
\usage{
registration(check = FALSE,
  myurl = "http://stephenturner.us/files/Tmisc-registration-link.txt")
}
\arguments{
\item{check}{logical; set to TRUE to check the registration link}

\item{myurl}{the URL to visit to print a registration link}
}
\value{
A registration link
}
\description{
Returns a link to register for a course. This reads a text file that I keep 
on my website that reveals a URL used to register for a course. I change the 
URL on my website, and the result of running the function changes. I use this
so that I can give instructions needed to set up for a course, which includes
instructions to download this package and run this function. This is used to
ensure that participants were able to successfully download and install R and
the required packages before coming to the course.
}
\examples{
\dontrun{
registration()
}

}
\author{
Stephen Turner
}
\keyword{NA}
