% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/safesave.R
\name{safesave}
\alias{safesave}
\title{safesave}
\usage{
safesave(...)
}
\arguments{
\item{...}{Arguments passed to `save()`}

\item{.whichop}{name of option to reference to determine if file should be saved. Recommended that this is unique to your file. Option should evaluate to logical. Safesave can change this option on exit, so don't do something dumb like set this to the name of a base option.}

\item{.keepopts}{If safesave should save, should it keep the .whichop option turned on for future calls to safesave? False by default. A single call to safesave, if .whichop is set to T, will switch it to FALSE after saving. Turn this option to True, to prevent safesave from changing the value of .whichop. This is useful if you want to set one option and allow safesave to run multiple times (as in the example below).}

\item{.format}{Save to rds or rda? Set to rda by default. If set to RDS, will set to RDS format. Setting format to RDS will require that no more than 1 object be specified for saving (this is a restriction fo the RDS format itself).}
}
\value{
prints message indicating whether save occured or how to set options to enable save
}
\description{
A function that only saves when options('saveopts') is set to TRUE
}
\details{

}
\examples{
\dontrun{
x<-rnorm(100)
options('saveopts'=F)
safesave(x,file='./tmp.Rda')
# will not save file
options('saveopts'=T)
safesave(x,file='./tmp.Rda')
# will save

# use case: setup a script and source, when you want to save it, set options before sourcing:
scriptdir<-tempdir()
script<-file.path(scriptdir,'script')
sink(script)
cat("x<-rnorm(100)\\n")
cat("y<-runif(199)\\n")
cat("z<-rpois(100,1)\\n")
cat("safesave(x,file=file.path(scriptdir,'tmp.Rda').whichop='saveopts',.keepopts=T)\\n")
cat("safesave(y,file=file.path(scriptdir,'tmp1.Rda'),.whichop='saveopts',.keepopts=F)\\n")
cat("safesave(z,file=file.path(scriptdir,'tmp2.Rda'))\\n")
sink()
#---
options('saveopts'=F) # default,
source('./script.R') # will run but not save
length(dir(scriptdir,pattern="tmp.*\\\\.Rda"))
# Script can save up to 3 files, but we don't always want it to do that. By default no files will save. to save (ex. from bash):
options('saveopts'=T)
source(script) # will save all three files.
dir(scriptdir,pattern="tmp.*\\\\.Rda")

#cleanup files created for this example
unlink(file.path(recursive=T,scriptdir))
# alternative in bash
# $ Rscript -e 'options("saveopts"=T);source(script)'
}
}

