% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveit.R, R/utils.R
\name{saveit}
\alias{saveit}
\title{Rename objects while saving.}
\usage{
saveit(..., file = stop("'file' must be specified"))

saveit(..., file = stop("'file' must be specified"))
}
\arguments{
\item{...}{Objects to save.}

\item{file}{Filename/path where data will be saved.}
}
\description{
Allows you to rename objects as you save them. See \url{https://stackoverflow.com/a/21248218/654296}.

Allows you to rename objects as you save them. See \url{https://stackoverflow.com/a/21248218/654296}.
}
\examples{
\dontrun{
foo <- 1
saveit(bar=foo, file="foobar.Rdata")
}
  
\dontrun{
foo <- 1
saveit(bar=foo, file="foobar.Rdata")
}
  
}
