% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_build.r
\docType{package}
\name{tmisc-package}
\alias{tmisc-package}
\title{This is just a few functions that are useful for me in writing my dissertation and doing daily analysis}
\description{
This is just a few functions that are useful for me in writing my dissertation and doing daily analysis
}

