\docType{package}
\name{Tmisc Package}
\alias{Tmisc Package}
\alias{Tmisc Package-package}
\title{This is just a few functions that are useful for me in writing my dissertation and doing daily analysis}
\description{
This is just a few functions that are useful for me in
writing my dissertation and doing daily analysis
}

