% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfactor.R
\name{unfactor}
\alias{unfactor}
\title{Unfactor a data.frame}
\usage{
unfactor(df)
}
\arguments{
\item{df}{The dataframe you wish to change the factors into characters.}
}
\value{
A data.frame with factors converted to characters.
}
\description{
Did you forget to pass \code{stringsAsFactors=FALSE}? This converts factor
variables to characters in a dataframe.
}
\examples{
df <- data.frame(a = letters[1:5], x = 1:5, y = LETTERS[1:5], stringsAsFactors = TRUE)
str(df)
df <- unfactor(df)
str(df)
  
}
\author{
\url{https://github.com/Dasonk}
}
\keyword{NA}
