% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unrowname.R
\name{unrowname}
\alias{unrowname}
\title{Un-rowname}
\usage{
unrowname(x)
}
\arguments{
\item{x}{A data.frame.}
}
\value{
A data.frame with rownames removed.
}
\description{
Strip rownames from an object (stolen from plyr).
}
\examples{
unrowname(mtcars)
  
}
\author{
Hadley Wickham
}
\keyword{keywords}

